+++ packages/interpreter/src/common.js
