+++ packages/interpreter/src/lib.rs
@@ -15,7 +15,7 @@ pub use bindings::Interpreter;
 #[cfg(feature = "minimal_bindings")]
 pub mod minimal_bindings {
     use wasm_bindgen::{prelude::wasm_bindgen, JsValue};
-    #[wasm_bindgen(module = "/src/setAttributeInner.js")]
+    #[wasm_bindgen(module = "/src/common.js")]
     extern "C" {
         pub fn setAttributeInner(node: JsValue, name: &str, value: JsValue, ns: Option<&str>);
     }
